backend/app/services/visual_analysis.py
import logging
from pathlib import Path
import cv2
import numpy as np
from typing import Dict, Any, Tuple, Optional, List
from pyzbar.pyzbar import decode as pyzbar_decode # For barcode detection

from app.core.config import settings
from app.utils import image_utils
from app.utils.custom_exceptions import VisualAnalysisError
from app.api.v1 import schemas as api_schemas

logger = logging.getLogger(settings.APP_NAME)

DEFAULT_DPI = settings.DEFAULT_IMAGE_DPI

def pixels_to_mm(pixels: float, dpi: int) -> float:
    if dpi <= 0: return 0.0
    return (pixels / dpi) * 25.4

# ... (other conversion functions like mm_to_pixels, etc., remain the same) ...

async def measure_font_size(
    # ... (function content remains the same as before) ...
    # This function is already designed to be called by the rule engine.
) -> Dict[str, Any]:
    # Placeholder for existing font size logic
    return {"status": "wrong", "message": "Font size check needs full implementation."}


async def detect_and_measure_barcode(
    image_path: Path,
    image_dpi: int = DEFAULT_DPI
) -> List[Dict[str, Any]]:
    """
    Detects all barcodes in an image and measures their dimensions.
    Args:
        image_path: Path to the label image.
        image_dpi: DPI of the image for accurate conversion.
    Returns:
        A list of dictionaries, each containing data for a detected barcode.
    """
    logger.info(f"Detecting and measuring barcodes in image: {image_path}")
    results = []
    try:
        img = await image_utils.load_image(image_path)
        # pyzbar works best with grayscale images
        gray_img = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
        
        detected_barcodes = pyzbar_decode(gray_img)
        
        if not detected_barcodes:
            logger.warning(f"No barcodes found in image {image_path.name}")
            return []

        for barcode in detected_barcodes:
            (x, y, w, h) = barcode.rect
            measured_width_mm = pixels_to_mm(w, image_dpi)
            measured_height_mm = pixels_to_mm(h, image_dpi)
            
            barcode_data = {
                "data": barcode.data.decode("utf-8"),
                "type": barcode.type,
                "bounding_box": api_schemas.BoundingBox(x=x, y=y, width=w, height=h),
                "measured_width_mm": measured_width_mm,
                "measured_height_mm": measured_height_mm,
            }
            results.append(barcode_data)
            logger.info(f"Detected barcode: {barcode_data}")

        return results
    except Exception as e:
        logger.error(f"Error during barcode detection: {e}", exc_info=True)
        raise VisualAnalysisError(f"An unexpected error occurred during barcode detection: {e}")

# ... (other visual check functions) ...
