backend/pyproject.toml
[tool.poetry]
name = "label-ai-backend"
version = "0.1.0"
description = "Backend for the AI-powered Regulatory Label Checker"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"

# Core FastAPI and related libraries
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
python-multipart = "^0.0.9"

# AI, CV, and File Parsing Libraries
pytesseract = "^0.3.10"
Pillow = "^10.3.0"
opencv-python-headless = "^4.9.0"
PyYAML = "^6.0.1"
python-docx = "^1.1.2"  # <-- ADDED: For reading .docx files
pyzbar = "^0.1.9"        # <-- ADDED: For decoding barcodes

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"

# Linting and Formatting
ruff = "^0.4.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# --- Tool Configurations ---
# (Ruff, Pytest configurations remain the same)
[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "UP", "C4", "B", "A", "COM", "LOG", "Q", "PT", "SIM", "TID", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

